requirements.txt
pandas
requests
google-cloud-storage
google-cloud-bigquery
apache-airflow
